Provide default user agent in BrowserContext
